+++ packages/localizations/src/es-MX.ts
@@ -154,6 +154,7 @@ export const esMX: LocalizationResource = {
     totalDue: undefined,
     totalDueToday: undefined,
     viewFeatures: undefined,
+    viewPayment: undefined,
     year: undefined,
   },
   createOrganization: {