+++ packages/localizations/src/sk-SK.ts
@@ -154,6 +154,7 @@ export const skSK: LocalizationResource = {
     totalDue: undefined,
     totalDueToday: undefined,
     viewFeatures: undefined,
+    viewPayment: undefined,
     year: undefined,
   },
   createOrganization: {