+++ packages/localizations/src/sr-RS.ts
@@ -154,6 +154,7 @@ export const srRS: LocalizationResource = {
     totalDue: undefined,
     totalDueToday: undefined,
     viewFeatures: undefined,
+    viewPayment: undefined,
     year: undefined,
   },
   createOrganization: {